some ui chnages

+++ src/services/authServices.js
@@ -21,28 +21,32 @@ const register = async (user) => {
   }
 };
 
-const login = async (username, password) => {
+const login = async (usernameOrEmail, password) => {
   try {
     const { data } = await axios.post("https://examination-system-backend-production.up.railway.app/api/login", {
-      username: username,
-      password: password,
+      username: usernameOrEmail,
+      password,
     });
 
-    if (data && data.jwtToken) {  // Check if jwtToken exists
+    if (data && data.jwtToken && data.user) {  // Check if both jwtToken and user exist
       localStorage.setItem("user", JSON.stringify(data.user));
       localStorage.setItem("jwtToken", JSON.stringify(data.jwtToken));
       console.log("authService:login() Success: ", data.user);
       return data;
     } else {
-      console.error("authService:login() Error: No JWT token found", data);
-      return { error: "No JWT token found" };  // Handle missing token
+      console.error("authService:login() Error: Missing data in response", data);
+      return { error: "Invalid response data" };  // Handle missing data
     }
   } catch (error) {
-    const errorMsg = error.response ? error.response.statusText : error.message;
+    // Improved error handling
+    const errorMsg = error.response
+      ? `${error.response.status} - ${error.response.statusText}`
+      : error.message;
     console.error("authService:login() Error: ", errorMsg);
     return { error: errorMsg };
   }
 };
+
 
 const authServices = { register, login };
 export default authServices;
